requirements.txt
astroid==2.2.5
colorama==0.4.1
entrypoints==0.3
flake8==3.7.7
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.16.4
pycodestyle==2.5.0
pyflakes==2.1.1
pygame==1.9.6
pylint==2.3.1
PyOpenGL==3.1.0
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
